refactor(cli): extract dashboard management into broker-dashboard.ts
Move the ~550-line dashboard cluster (static-asset discovery, relay-URL
resolution, spawn env/args, process startup, port readiness, and stale UI
asset refresh) out of broker-lifecycle.ts into a dedicated broker-dashboard.ts.

broker-lifecycle.ts now imports a 7-function dashboard boundary
(resolveDashboardPortWithFallback, resolveDashboardRelayUrl,
getDefaultDashboardRelayUrl, isDebugLikeLoggingEnabled, normalizeDashboardPath,
startDashboardWithFallback, waitForDashboard) and keeps everything else private
to the dashboard module. broker-lifecycle.ts drops from 1879 to ~1320 lines.

Pure move with no behavior change. Also collapses a few multi-line imports in
the step-1 mcp/ files via prettier.
